changed return type of load_text_file() and read_text_file from int to size_t in FileUtils.c

+++ src/FileUtils.h
@@ -2,6 +2,7 @@
 
 #include <stdio.h>
 #include <stdbool.h>
+#include <limits.h>
 
 
 
@@ -21,8 +22,8 @@ void build_path(char *path, const size_t num, ...);
 int get_lock(char *lock_file, const bool blocking);
 void free_lock(char *lock_file, const int handle);
 
-int load_text_file(const char *file, char **out);
-int read_text_file(FILE *handle, char **out);
-bool save_text_file(const char *file, char *data, size_t size);
+size_t load_text_file(const char *file, char **out);
+size_t read_text_file(FILE *handle, char **out);
+bool   save_text_file(const char *file, char *data, size_t size);
 
 bool copy_file(char *file_path, char *dest_path);
